fix(ssr): handle symboless Modules in dynamic imports
