+++ packages/router/src/utils/index.ts
@@ -3,12 +3,36 @@ import {
   RouteComponent,
   RouteParamsRawGeneric,
   RouteParamValueRaw,
+  RawRouteComponent,
 } from '../types'
 
 export * from './env'
 
+/**
+ * Allows differentiating lazy components from functional components and vue-class-component
+ * @internal
+ *
+ * @param component
+ */
+export function isRouteComponent(
+  component: RawRouteComponent
+): component is RouteComponent {
+  return (
+    typeof component === 'object' ||
+    'displayName' in component ||
+    'props' in component ||
+    '__vccOpts' in component
+  )
+}
+
 export function isESModule(obj: any): obj is { default: RouteComponent } {
-  return obj.__esModule || obj[Symbol.toStringTag] === 'Module'
+  return (
+    obj.__esModule ||
+    obj[Symbol.toStringTag] === 'Module' ||
+    // support CF with dynamic imports that do not
+    // add the Module string tag
+    (obj.default && isRouteComponent(obj.default))
+  )
 }
 
 export const assign = Object.assign
